Able to randomize deck.
Example: /random/Lesson%201 will randomize the cards in Lesson 1.

Clearly, this causes a conflict with randomize by word, like /random/concise.
In this case, deck will always have higher priority, that is, if there is a deck
called "Lesson 1" and a card with its word set to "Lesson 1", /random/Lesson%201
will randomize the deck.

The rationale behind this design is that words can be selected using ID, but
decks cannot. Hence, if such conflict arrises, to select the card, just use
/random/<id>, where id is the ID of the card.